Refactor CohortManager: move selected to own file, update Cohort Panel
