+++ packages/core/src/features/cohort/reducers.ts
@@ -1,5 +1,5 @@
 import { combineReducers } from '@reduxjs/toolkit';
-import { cohortReducer } from './cohortSlice';
+import { cohortReducer } from './cohortManagerSlice';
 import { cohortBuilderFiltersExpandedReducer } from './filterExpandSlice';
 import { cohortBuilderFiltersCombineModeReducer } from './filterCombineModeSlice';
 import { cohortSharedFiltersReducer } from './sharedFiltersSlice';
@@ -8,5 +8,5 @@ export const cohortReducers = combineReducers({
   filtersExpanded: cohortBuilderFiltersExpandedReducer,
   filtersCombineMode: cohortBuilderFiltersCombineModeReducer,
   sharedFilters: cohortSharedFiltersReducer,
-  cohort: cohortReducer,
+  cohortManager: cohortReducer,
 });
