+++ packages/core/src/index.ts
@@ -1,19 +1,19 @@
 import {
+  Accessibility,
   GEN3_API,
-  GEN3_DOMAIN,
-  GEN3_COMMONS_NAME,
-  GEN3_DOWNLOADS_ENDPOINT,
-  GEN3_GUPPY_API,
-  GEN3_FENCE_API,
   GEN3_AUTHZ_API,
+  GEN3_COMMONS_NAME,
+  GEN3_CROSSWALK_API,
+  GEN3_DOMAIN,
+  GEN3_DOWNLOADS_ENDPOINT,
+  GEN3_FENCE_API,
+  GEN3_GUPPY_API,
+  GEN3_MANIFEST_API,
   GEN3_MDS_API,
   GEN3_REDIRECT_URL,
-  GEN3_WORKSPACE_API,
-  GEN3_SUBMISSION_API,
-  GEN3_CROSSWALK_API,
   GEN3_SOWER_API,
-  GEN3_MANIFEST_API,
-  Accessibility,
+  GEN3_SUBMISSION_API,
+  GEN3_WORKSPACE_API,
 } from './constants';
 import { type CoreState } from './reducers';
