--- packages/frontend/src/features/CohortBuilder/CohortManager.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import QueryExpression from './QueryExpression';
-
-interface CohortManagerProps {
-  index: string;
-}
-
-const CohortManager = ({ index }: CohortManagerProps) => {
-  return (
-    <div className="mb-2">
-      <QueryExpression index={index} />
-    </div>
-  );
-};
-
-export default CohortManager;
