+++ packages/frontend/src/features/CohortBuilder/CohortManagerAndExpression.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import QueryExpression from './QueryExpression';
+import CohortManager from './CohortManager/CohortManager';
+
+interface CohortManagerProps {
+  index: string;
+}
+
+const CohortManagerAndExpression = ({ index }: CohortManagerProps) => {
+  return (
+    <div className="flex flex-col mb-2">
+      <CohortManager />
+      <QueryExpression index={index} />
+    </div>
+  );
+};
+
+export default CohortManagerAndExpression;
